requirements.txt
Flask
Flask-Script
Flask-PyMongo
flask-restful
itsdangerous
